Fixed checkout, updated cart, and updated article page as well
Added new feature in shopping cart. Now it points out the item listed in cart are available on selected store and online store or not.
In checkout fixed some styling and also fixed item removal so order products are ok.
Also made order no dynamic.
In product page added new feature if the product is available in the store it wont open the dialog box